requirements.txt
modal>=1.2.1
torch
torchvision
diffusers[torch]
transformers>=4.25.1
ftfy
Pillow
oyaml
opencv-python
albumentations
safetensors
lycoris-lora==1.8.3
flatten_json
pyyaml
tensorboard
kornia
invisible-watermark>=0.1.5
einops
accelerate
toml
pydantic
omegaconf
python-dotenv==1.0.1
k-diffusion
open_clip_torch
timm
prodigyopt
controlnet_aux==0.0.7
bitsandbytes
hf_transfer
lpips
pytorch_fid
optimum-quanto
sentencepiece
huggingface_hub
peft
